Fixed the bounding box's resize method to accept doubles

+++ src/sherwood/demo/physics/BoundingBox.java
@@ -56,7 +56,7 @@ public class BoundingBox {
         return new Rectangle(position.xc(), position.yc(), size.xc(), size.yc());
     }
 
-    public BoundingBox resize (int width, int height) {
+    public BoundingBox resize (double width, double height) {
         return new BoundingBox(position, new Vector(width, height));
     }
 }
